move all siteSettings related queries to siteSetting.ts
