+++ src/app/(main)/layout.tsx
@@ -13,9 +13,11 @@ import {
   COMPANY_INFO_QUERY,
   LEGAL_DOCUMENTS_BY_LANG_QUERY,
 } from "studio/lib/queries/admin";
-import { BRAND_ASSETS_QUERY } from "studio/lib/queries/brandAssets";
-import { NAV_QUERY } from "studio/lib/queries/navigation";
-import { SOMEPROFILES_QUERY } from "studio/lib/queries/socialMediaProfiles";
+import {
+  BRAND_ASSETS_QUERY,
+  NAV_QUERY,
+  SOME_PROFILES_QUERY,
+} from "studio/lib/queries/siteSettings";
 import { loadStudioQuery } from "studio/lib/store";
 
 import styles from "./layout.module.css";
@@ -39,7 +41,7 @@ export default async function Layout({
     loadStudioQuery<Navigation>(NAV_QUERY, {}, { perspective }),
     loadStudioQuery<CompanyInfo>(COMPANY_INFO_QUERY, {}, { perspective }),
     loadStudioQuery<SocialMediaProfiles | null>(
-      SOMEPROFILES_QUERY,
+      SOME_PROFILES_QUERY,
       {},
       { perspective },
     ),
