+++ src/app/sitemap.ts
@@ -4,7 +4,7 @@ import { client } from "studio/lib/client";
 import { DocumentWithSlug } from "studio/lib/interfaces/global";
 import { PageBuilder } from "studio/lib/interfaces/pages";
 import { DOCUMENTS_WITH_SLUG_QUERY } from "studio/lib/queries/document";
-import { LANDING_PAGE_QUERY } from "studio/lib/queries/navigation";
+import { LANDING_PAGE_QUERY } from "studio/lib/queries/siteSettings";
 import { token } from "studio/lib/token";
 
 import { readBaseUrl } from "./env";
