+++ src/utils/seo.ts
@@ -8,8 +8,10 @@ import { BrandAssets } from "studio/lib/interfaces/brandAssets";
 import { CompanyInfo } from "studio/lib/interfaces/companyDetails";
 import { DefaultSeo } from "studio/lib/interfaces/defaultSeo";
 import { COMPANY_INFO_QUERY } from "studio/lib/queries/admin";
-import { BRAND_ASSETS_QUERY } from "studio/lib/queries/brandAssets";
-import { DEFAULT_SEO_QUERY } from "studio/lib/queries/seo";
+import {
+  BRAND_ASSETS_QUERY,
+  DEFAULT_SEO_QUERY,
+} from "studio/lib/queries/siteSettings";
 import { loadStudioQuery } from "studio/lib/store";
 
 type SeoData = {
