--- studio/lib/queries/languages.ts
@@ -1,3 +1,0 @@
-import { groq } from "next-sanity";
-
-export const DEFAULT_LANGUAGE_QUERY = groq`*[_type == "languageSettings" && _id == "languageSettings"][0].languages[default][0]`;
