--- studio/lib/queries/socialMediaProfiles.ts
@@ -1,5 +1,0 @@
-import { groq } from "next-sanity";
-
-export const SOMEPROFILES_QUERY = groq`
-  *[_type == "soMeLinksID" && _id == "soMeLinksID"][0]
-`;
