+++ studio/schemas/objects/footerSection.ts
@@ -2,7 +2,7 @@ import { StringInputProps, defineType } from "sanity";
 
 import { StringInputWithCharacterCount } from "studio/components/stringInputWithCharacterCount/StringInputWithCharacterCount";
 import { client } from "studio/lib/client";
-import { SOMEPROFILES_QUERY } from "studio/lib/queries/socialMediaProfiles";
+import { SOME_PROFILES_QUERY } from "studio/lib/queries/siteSettings";
 import { soMeLinksID } from "studio/schemas/documents/siteSettings/socialMediaProfiles";
 import { richText } from "studio/schemas/fields/text";
 
@@ -90,7 +90,7 @@ export const footerSection = defineType({
       hidden: ({ parent }) => parent?.sectionType !== SectionType.SocialMedia,
       initialValue: async () => {
         // use Social Media Profiles singleton document if it exists
-        return (await client.fetch(SOMEPROFILES_QUERY)) !== null
+        return (await client.fetch(SOME_PROFILES_QUERY)) !== null
           ? {
               _type: "reference",
               _ref: soMeLinksID,
